Bullet Area base
- Sostituito l'OverlapCapsule con un OverlapBox per una migliore gestione delle collisioni;
- Itegrate variabili per l'area e il multiplier in vista degli upgrade
- Integrate despawn e movimento in un unico script per il bullet ad Area

ps: possibile ottimizzazione: prendere il codice di "OverlapDam.cs" e integrarlo nello script dei bullet, e vedere di modificare il colore dei bullet da script in base al colore delle torrette;

+++ Assets/Scripts/EnemyThings/Enemy.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class Enemy : MonoBehaviour, IDamageable
 {
     [Header("Health settings")]
-    [SerializeField] float MaxHealth;
-    public float CurrentHealth;
+    [SerializeField] int MaxHealth;
+    public int CurrentHealth;
 
     public static event System.Action<int> OnEnemyKilled;
